packages/LeRobot_AgiBotWorldChallenge_2025/requirements.txt
action-msgs==1.2.1
actionlib-msgs==4.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
ament-cmake-test==1.3.11
ament-copyright==0.12.12
ament-cppcheck==0.12.12
ament-cpplint==0.12.12
ament-flake8==0.12.12
ament-index-python==1.4.0
ament-lint==0.12.12
ament-lint-cmake==0.12.12
ament-package==0.14.1
ament-pep257==0.12.12
ament-uncrustify==0.12.12
ament-xmllint==0.12.12
annotated-types==0.7.0
argcomplete==1.8.1
async-timeout==5.0.1
attrs==25.3.0
av==15.0.0
blinker==1.9.0
bloom==0.13.0
builtin-interfaces==1.2.1
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
cmake==4.0.3
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.2.1
colcon-cmake==0.2.29
colcon-common-extensions==0.3.0
colcon-core==0.19.0
colcon-defaults==0.2.9
colcon-devtools==0.3.0
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-mixin==0.2.3
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colcon-zsh==0.5.0
composition-interfaces==1.2.1
cryptography==3.4.8
cv-bridge==3.2.1
datasets==3.6.0
dbus-python==1.2.18
deepdiff==8.6.0
diagnostic-msgs==4.9.0
diffusers==0.34.0
dill==0.3.8
distlib==0.3.4
distro==1.7.0
docutils==0.17.1
domain-coordinator==0.10.0
draccus==0.10.0
einops==0.8.1
empy==3.3.4
evdev==1.9.2
Farama-Notifications==0.0.4
filelock==3.18.0
flake8==4.0.1
Flask==3.1.1
frozenlist==1.7.0
fsspec==2025.3.0
geometry-msgs==4.9.0
gitdb==4.0.12
GitPython==3.1.45
gymnasium==0.29.1
hf-xet==1.1.7
hf_transfer==0.1.9
httplib2==0.20.2
huggingface-hub==0.33.5
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib-metadata==4.6.4
iniconfig==1.1.1
inquirerpy==0.3.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
jsonlines==4.0.0
keyring==23.5.0
lark==1.1.1
launch==1.0.9
launch-ros==0.19.9
launch-testing==1.0.9
launch-testing-ros==0.19.9
launch-xml==1.0.9
launch-yaml==1.0.9
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
# Editable Git install with no remote (lerobot==0.2.0)
-e /root/workspace/main/AgiBot-World
lifecycle-msgs==1.2.1
lxml==4.8.0
MarkupSafe==3.0.2
mccabe==0.6.1
mergedeep==1.3.4
message-filters==4.3.7
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
nav-msgs==4.9.0
netifaces==0.11.0
networkx==3.4.2
notify2==0.3
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.2.0
opencv-python==4.7.0.72
opencv-python-headless==4.12.0.88
orderly-set==5.5.0
osrf-pycommon==2.1.6
packaging==25.0
pandas==2.3.1
pfzy==0.3.4
pillow==11.3.0
platformdirs==4.3.8
pluggy==0.13.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==5.9.0
py==1.10.0
pyarrow==21.0.0
pycodestyle==2.8.0
pydantic==2.11.7
pydantic_core==2.33.2
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pynput==1.8.1
pyparsing==2.4.7
pyserial==3.5
pytest==6.2.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-xlib==0.33
pytz==2025.2
PyYAML==6.0.2
pyyaml-include==1.4.1
rcl-interfaces==1.2.1
rclpy==3.3.16
rcutils==5.1.6
regex==2025.7.34
requests==2.32.4
rerun-sdk==0.22.1
rmw-dds-common==1.6.0
roman==3.3
ros2action==0.18.12
ros2bag==0.15.14
ros2cli==0.18.12
ros2component==0.18.12
ros2doctor==0.18.12
ros2interface==0.18.12
ros2launch==0.19.9
ros2lifecycle==0.18.12
ros2multicast==0.18.12
ros2node==0.18.12
ros2param==0.18.12
ros2pkg==0.18.12
ros2run==0.18.12
ros2service==0.18.12
ros2topic==0.18.12
rosbag2-interfaces==0.15.14
rosbag2-py==0.15.14
rosdep==0.26.0
rosdep-modules==0.26.0
rosdistro==1.0.1
rosdistro-modules==1.0.1
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.6
rosidl-cli==3.1.6
rosidl-cmake==3.1.6
rosidl-generator-c==3.1.6
rosidl-generator-cpp==3.1.6
rosidl-generator-py==0.14.4
rosidl-parser==3.1.6
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.6
rosidl-typesupport-introspection-cpp==3.1.6
rospkg-modules==1.6.0
rpyutils==0.2.1
safetensors==0.6.2
SecretStorage==3.3.1
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
sentry-sdk==2.34.1
shape-msgs==4.9.0
six==1.16.0
smmap==5.0.2
snowballstemmer==2.2.0
sros2==0.10.6
statistics-msgs==1.2.1
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
sympy==1.14.0
termcolor==3.1.0
tf2-geometry-msgs==0.25.13
tf2-kdl==0.25.13
tf2-msgs==0.25.13
tf2-py==0.25.13
tf2-ros-py==0.25.13
tf2-sensor-msgs==0.25.13
tf2-tools==0.25.13
theora-image-transport==2.5.3
tokenizers==0.21.4
toml==0.10.2
torch==2.8.0
torchcodec==0.6.0
torchvision==0.23.0
tqdm==4.67.1
trajectory-msgs==4.9.0
transformers==4.51.3
triton==3.4.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unique-identifier-msgs==2.2.1
urllib3==2.5.0
vcstool==0.3.0
vcstools==0.1.42
visualization-msgs==4.9.0
wadllib==1.3.6
wandb==0.21.1
wcwidth==0.2.13
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
zipp==1.0.0
